Add requirement of trnascan-se > 2, handle output when no trnas detected

+++ mag_annotator/annotate_bins.py
@@ -6,6 +6,7 @@ import pandas as pd
 from datetime import datetime
 import re
 from glob import glob
+import warnings
 
 from mag_annotator.utils import run_process, make_mmseqs_db, merge_files, get_database_locs
 
@@ -333,9 +334,12 @@ def run_trna_scan(fasta, output_loc, fasta_name, threads=10, verbose=True):
     raw_trnas = path.join(output_loc, 'raw_trnas.txt')
     run_process(['tRNAscan-SE', '-G', '-o', raw_trnas, '--thread', str(threads), fasta], verbose=verbose)
     processed_trnas = path.join(output_loc, 'trnas.tsv')
-    trna_frame = pd.read_csv(raw_trnas, sep='\t', skiprows=[0, 2], index_col=0)
-    trna_frame.insert(0, 'fasta', fasta_name)
-    trna_frame.to_csv(processed_trnas, sep='\t')
+    if path.isfile(processed_trnas):
+        trna_frame = pd.read_csv(raw_trnas, sep='\t', skiprows=[0, 2], index_col=0)
+        trna_frame.insert(0, 'fasta', fasta_name)
+        trna_frame.to_csv(processed_trnas, sep='\t')
+    else:
+        warnings.warn('No tRNAs were detected, no trnas.tsv file will be created.')
 
 
 def do_blast_style_search(query_db, target_db, working_dir, header_dict, get_description, start_time,
